Change project build structure (#1)
Well, as the base project is now archived I do not need to try to be
compatible with it.
The main reason behind this changes is to simplify overall project
building (and make it build-able again).
Current drawbacks:
* There is no sound (fmod was on now-dead conan repo and I think it's
the time to migrate to openal)
* mimalloc is removed (I have heap corruption when it's linked so I
decided to remove it for now)
* There are no modules anymore - all sources are compiled directly to
executable. This is planned as a temporary solution, but even if I
return the modules back, they will be different. I've done this becase
old modules state is just... meh. They are so tightly bind together that
it's simpler to just merge them all to executable and then step by step
separate independent code to the new modules.
* CMake Multi-generators are not supported and there is no plans to
support them anymore