Go back to using INI for configs (#13)
Well, I've changed my mind. There is no many reasons to switch to another config formats.
INI configs are readable and works fine. They need a custom parser, but
it is not much of a problem.

So I wrote a new parser - much simpler and more readable than the old
one. For now it's only engine.ini and aliases uses are adapted
for the new parser. Aliases files itself are remain unchanged, so no
converting required.
engine.ini refactored a bit and some key names are renamed and moved to
specific sections (this is not final and additional adjustments are
possible)